+++ src/components/Navigation/TopBar/Help/FeedbackForm/UploadFile.tsx
@@ -4,7 +4,7 @@ import { FileRejection, FileWithPath } from 'react-dropzone';
 import { Typography } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 
-import { FeedbackContentType } from './FeedbackForm';
+import { FeedbackContentType } from './FeedbackForm.types';
 import { SeverityOption } from './FeedbackFormInner';
 import FileProgress from 'src/components/Feedback/Progress/FileProgress';
 import FileUploadArea from 'src/components/Inputs/FileUploadArea';
